allow plan steps to always be executed
