+++ pkg/plan/step_condition.go
@@ -1,0 +1,7 @@
+package plan
+
+const (
+	ConditionAlwaysExecute = "always-execute"
+)
+
+type StepCondition string
